+++ fflas-ffpack/ffpack/ffpack_permutation.inl
@@ -471,7 +471,7 @@ namespace FFPACK {
 	papplyP( const Field& F,
 		 const FFLAS::FFLAS_SIDE Side,
 		 const FFLAS::FFLAS_TRANSPOSE Trans,
-		 const size_t m, const int ibeg, const int iend,
+		 const size_t m, const size_t ibeg, const size_t iend,
 		 typename Field::Element_ptr A, const size_t lda, const size_t * P )
 	{
 		int numthreads = omp_get_max_threads();
